Update oauth.js
adding support for pdf responses. This is a hack that works for me, I am not sure if this is the optimal solution as I am not sure if PDF are binary or not (if they were binary it should fall I guest in the first case of the large if then else.